Add files via upload
This project is a responsive e-commerce website designed to replicate the shopping experience of Amazon. It features a visually appealing layout with a navigation bar that allows users to access various sections such as account information, search functionality, and delivery location. The homepage includes promotional messages and product categories, each linked to more detailed listings, making it easy for users to explore different items. The design incorporates a clean aesthetic with interactive elements that enhance user engagement. Overall, the project aims to provide a seamless and user-friendly online shopping experience.